+++ source/getentropy_getrandom.c
@@ -23,7 +23,7 @@
 #include <sys/random.h>
 #include <errno.h>
 
-int xp_getentropy(void* buf, size_t len) {
+int xp_getentropy(void* buf, uint32_t len) {
     if (len > 256) {
         errno = EIO;
         return -1;
@@ -35,7 +35,7 @@ int xp_getentropy(void* buf, size_t len) {
     do {
         ret = getrandom(buf, len, GRND_NONBLOCK);
     } while (ret < 0 && errno == EINTR);
-    if ((size_t)ret != len) {
+    if (ret != len) {
         // May be EAGAIN, EINVAL, ENOSYS, etc but getentropy can only have 1 of 2 errnos.
         if (errno != EFAULT) errno = EIO;
         return -1;
