+++ source/getentropy_sysctl.c
@@ -45,7 +45,7 @@ static inline size_t getentropy_sysctl(u_char* buf, size_t size) {
     return done;
 }
 
-int xp_getentropy(void* buf, size_t len) {
+int xp_getentropy(void* buf, uint32_t len) {
     if (len > 256) {
         errno = EIO;
         return -1;
